+++ docs/changes.rst
@@ -6,6 +6,8 @@ weblate 3.2
 
 Released on ? 2018.
 
+* Add install_addon management command for automated addon installation.
+
 weblate 3.1.1
 -------------